style: fix Rust 1.65 clippy lints (#335)

+++ hal-core/src/addr.rs
@@ -32,7 +32,7 @@ pub trait Address:
             return self;
         }
         let aligned = (u | mask) + 1;
-        Self::from_usize(aligned as usize)
+        Self::from_usize(aligned)
     }
 
     /// Align `self` up to the required alignment for a value of type `T`.
